+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add `all | qualified` toggle to Stats and save sort + column preferences in
+  memory: [PR 137](https://github.com/mlb-rs/mlbt/pull/137)
+ 
 ## [0.3.1] - 2026-04-16
 
 ### Fixed